+++ tests/Post/FunctionsTest.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Fire\Tests\Post;
 
 use Fire\Post\Type\ArchivePageSetting;
+use Fire\Post\Type\Support;
 use Fire\Tests\TestCase;
 
 use function Brain\Monkey\Functions\expect;
+use function Brain\Monkey\Functions\when;
 use function Fire\Post\id;
 use function Fire\Post\page_id_for_type;
 
@@ -18,30 +20,77 @@ final class FunctionsTest extends TestCase
         $type = 'custom';
         $id = 1;
 
+        when('is_home')->justReturn(false);
+
+        expect('post_type_supports')
+            ->once()
+            ->with($type, Support::ARCHIVE_PAGE)
+            ->andReturn(true);
+
         expect('get_option')
             ->once()
             ->with(ArchivePageSetting::optionName($type))
             ->andReturn($id);
 
-        $this->assertSame(
-            $id,
-            page_id_for_type($type)
-        );
+        $this->assertSame($id, page_id_for_type($type));
     }
 
     public function testPageIdForTypeWithPosts(): void
     {
         $id = 1;
 
+        when('is_home')->justReturn(false);
+
+        expect('post_type_supports')
+            ->once()
+            ->with('post', Support::ARCHIVE_PAGE)
+            ->andReturn(true);
+
         expect('get_option')
             ->once()
             ->with(ArchivePageSetting::optionName('posts'))
             ->andReturn($id);
 
-        $this->assertSame(
-            $id,
-            page_id_for_type('post')
-        );
+        $this->assertSame($id, page_id_for_type('post'));
+    }
+
+    public function testPageIdForTypeIsHome(): void
+    {
+        $id = 1;
+
+        when('is_home')->justReturn(true);
+
+        expect('get_query_var')
+            ->once()
+            ->andReturn('');
+
+        expect('post_type_supports')
+            ->once()
+            ->with('post', Support::ARCHIVE_PAGE)
+            ->andReturn(true);
+
+        expect('get_option')
+            ->once()
+            ->with(ArchivePageSetting::optionName('posts'))
+            ->andReturn($id);
+
+        $this->assertSame($id, page_id_for_type());
+    }
+
+    public function testPageIdForTypeIsHomeDoesNotSupport(): void
+    {
+        when('is_home')->justReturn(true);
+
+        expect('get_query_var')
+            ->once()
+            ->andReturn('');
+
+        expect('post_type_supports')
+            ->once()
+            ->with('post', Support::ARCHIVE_PAGE)
+            ->andReturn(false);
+
+        $this->assertSame(0, page_id_for_type());
     }
 
     public function testId(): void
@@ -70,12 +119,17 @@ final class FunctionsTest extends TestCase
             ->andReturn(1);
 
         expect('is_home')
-            ->once()
+            ->twice()
             ->andReturn(true);
 
         expect('get_query_var')
             ->once()
-            ->andReturn('post');
+            ->andReturn('');
+
+        expect('post_type_supports')
+            ->once()
+            ->with('post', Support::ARCHIVE_PAGE)
+            ->andReturn(true);
 
         expect('get_option')
             ->once()